AKNodeRecorder playground
I finally trashed AKTapeRecorder. It worked fine in simulator and
playground, but crashes on hardware device. So I preferred to demo in a
simpler way, that really works in real world…  Playground page has been
renamed.
